feat: add custom hook for searchQuery
