+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\App.tsx":"1","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\ButtonError.tsx":"2"},{"size":1846,"mtime":1720392586220,"results":"3","hashOfConfig":"4"},{"size":515,"mtime":1720392630162,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","suppressedMessages":"8","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1i6yqbh",{"filePath":"9","messages":"10","suppressedMessages":"11","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\App.tsx",[],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\ButtonError.tsx",[],[]]
+[{"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\App.tsx":"1","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\ButtonError.tsx":"2","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\description-field\\DescriptionField.tsx":"3","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\loader\\Loading.tsx":"4","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\result\\Results.tsx":"5","C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\search\\Search.tsx":"6"},{"size":2002,"mtime":1720874630103,"results":"7","hashOfConfig":"8"},{"size":395,"mtime":1720862854147,"results":"9","hashOfConfig":"8"},{"size":348,"mtime":1720874108614,"results":"10","hashOfConfig":"8"},{"size":222,"mtime":1720873705722,"results":"11","hashOfConfig":"8"},{"size":1000,"mtime":1720873623710,"results":"12","hashOfConfig":"8"},{"size":734,"mtime":1720863036617,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","suppressedMessages":"16","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"1i6yqbh",{"filePath":"17","messages":"18","suppressedMessages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","suppressedMessages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","suppressedMessages":"25","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","suppressedMessages":"28","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","suppressedMessages":"31","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\App.tsx",["32"],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\ButtonError.tsx",[],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\description-field\\DescriptionField.tsx",[],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\loader\\Loading.tsx",[],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\result\\Results.tsx",[],[],"C:\\Users\\User\\Desktop\\react-rsschool\\class-components\\src\\components\\search\\Search.tsx",[],[],{"ruleId":"33","severity":1,"message":"34","line":36,"column":6,"nodeType":"35","endLine":36,"endColumn":14,"suggestions":"36"},"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'defaultValue'. Either include it or remove the dependency array. Outer scope values like 'SEARCH' aren't valid dependencies because mutating them doesn't re-render the component.","ArrayExpression",["37"],{"desc":"38","fix":"39"},"Update the dependencies array to be: [defaultValue]",{"range":"40","text":"41"},[1193,1201],"[defaultValue]"]
